Fixed data type used for pitchbend values across solution
Relates to #32
Updated all pitchbend functions to use the same data type (int16_t).
There was an issue where once place was still uint16_t causing it to use the wrong value.

Include a minor fix to IO logic.

+++ src/Common/PitchBend.h
@@ -17,7 +17,7 @@
 		PitchBend(){}; //Disallow creating an instance
 		
 		public:
-			inline static float calculateFactor(int bend, bool shiftRange = false)
+			inline static float calculateFactor(int16_t bend, bool shiftRange = false)
 			{
 				if(shiftRange) bend -= 8192;
 				float result = pow(2.0, (BEND_PER_OCTAVE * (bend / 8192.0)));
